refactor: move header outside files page
License: MIT
